Rename HistogramAlg.py to histogram_alg.py
