+++ test/recon_vis/test_utils.py
@@ -5,7 +5,7 @@ import shutil
 from empress import newickFormatReader
 from empress.recon_vis import utils
 from empress.reconcile import recongraph_tools
-from empress.histogram import HistogramBruteForce
+from empress.histogram import histogram_brute_force
 from empress.reconcile.script02_gen_newick_trees import generateNewickTestsMultipleSizes
 
 class TestUtils(unittest.TestCase):
@@ -239,7 +239,7 @@ class TestUtils(unittest.TestCase):
                 parasite_tree = recon_input.parasite_tree
                 for d, t, l in itertools.product(range(1, 5), repeat=3):
                     recon_graph, _, _, best_roots = recongraph_tools.DP(recon_input, d, t, l)
-                    for reconciliation, _ in HistogramBruteForce.BF_enumerate_MPRs(recon_graph, best_roots):
+                    for reconciliation, _ in histogram_brute_force.BF_enumerate_MPRs(recon_graph, best_roots):
                         temporal_graph = utils.build_temporal_graph(host_tree, parasite_tree, reconciliation)
                         ordering_dict = utils.topological_order(temporal_graph)
                         # the reconciliatin is strongly consistent
